% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrimMao.R
\name{predCrimMao}
\alias{predCrimMao}
\title{Spatio-temporal prediction of crimes in Manaus}
\usage{
predCrimMao(x, type = "ROUBO", nMap = 2, delta = 0.89, by1 = "week",
  k0 = NULL, UsePrior = FALSE, m0 = NULL, c0 = NULL)
}
\arguments{
\item{x}{A data frame containig the following columns:LAT,LONG,DATA,DIA.SEMANA,
HORA,PERIODO,Crime.}

\item{type}{Character string giving the type of crime to make the predictions. In this version only "ROUBO" is available.}

\item{nMap}{Number of maps to check the quality of predictions.}

\item{delta}{Discount factor.}

\item{by1}{Character string specifying the periodicity of the counts. In this version only "week" is available.}

\item{k0}{Number of periods to use as prior information.}

\item{UsePrior}{Default is \code{FALSE}, if \code{TRUE} the algorithm will take the informations
of the last prediction from \code{crimes1415} data set to use as prior. Incompatible with \code{k0}.}

\item{m0}{First posterior moment at t-1.}

\item{c0}{Second posterior moment at t-1.}
}
\value{
A map that provides predictions of crime intensity and the mean square error.
}
\description{
This function develops an one-step ahead prediction algorithm based on Poisson dynamic generalized
linear model (DGLM) for the count of crime in Manaus.
}
\examples{
data(crimes1415)
predCrimMao(crimes1415,type="ROUBO",by1="week",nMap=2)

}
\references{
Harrison, Jeff, and Mike West. \emph{Bayesian forecasting & dynamic models}. New York: Springer, 1999.

TRIANTAFYLLOPOULOS, K. Dynamic z
\emph{generalized linear models for non-Gaussian time series forecasting}. arXiv preprint arXiv:0802.0219, 2008.
}

