% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrimMaoWeb.R
\name{predCrimMaoWeb}
\alias{predCrimMaoWeb}
\title{Spatio-temporal prediction of crimes in Manaus for Web Visualization}
\usage{
predCrimMaoWeb(x, type = "ROUBO", delta = 0.89, by1 = "week", m0 = NULL,
  c0 = NULL, r.t0 = NULL, s.t0 = NULL, sigVg = NULL)
}
\arguments{
\item{x}{A data frame containig the following columns:LAT,LONG,DATA,DIA.SEMANA,
HORA,PERIODO,Crime.}

\item{type}{Character string giving the type of crime to make the predictions. In this version only "ROUBO" is available.}

\item{delta}{Discount factor.}

\item{by1}{Character string specifying the periodicity of the counts. In this version only "week" is available.}
}
\value{
A map that provides predictions of crime intensity and the mean square error.
}
\description{
This function develops an one-step ahead prediction algorithm based on Poisson dynamic generalized
linear model (DGLM) for the count of crime in Manaus.
}
\examples{
data(lastFiveWeek)
predCrimMaoWeb(lastFiveWeek,type="ROUBO",by1="week")

}
\references{
Harrison, Jeff, and Mike West. \emph{Bayesian forecasting & dynamic models}. New York: Springer, 1999.

TRIANTAFYLLOPOULOS, K. Dynamic z
\emph{generalized linear models for non-Gaussian time series forecasting}. arXiv preprint arXiv:0802.0219, 2008.
}

